+++ ocaml-lsp-server/src/ocaml_lsp_server.ml
@@ -1,6 +1,7 @@
 open Import
 module Version = Version
 module Diagnostics = Diagnostics
+module Position = Position
 module Doc_to_md = Doc_to_md
 module Diff = Diff
 module Testing = Testing
@@ -152,6 +153,7 @@ let initialize_info (client_capabilities : ClientCapabilities.t) :
       ~semanticTokensProvider
       ~experimental
       ~renameProvider
+      ~inlayHintProvider:(`Bool true)
       ~workspace
       ~executeCommandProvider
       ?positionEncoding
@@ -591,7 +593,8 @@ let on_request :
             Compl.resolve doc ci resolve Document.Merlin.doc_comment ~markdown))
       ()
   | CodeAction params -> Code_actions.compute server params
-  | InlayHint _ -> now None
+  | InlayHint params ->
+    later (fun state () -> Inlay_hints.compute state params) ()
   | TextDocumentColor _ -> now []
   | TextDocumentColorPresentation _ -> now []
   | TextDocumentHover req ->