+++ ocaml-lsp-server/src/ocaml_lsp_server.mli
@@ -2,5 +2,6 @@ val run : Lsp.Cli.Channel.t -> read_dot_merlin:bool -> unit -> unit
 
 module Diagnostics = Diagnostics
 module Version = Version
+module Position = Position
 module Doc_to_md = Doc_to_md
 module Testing = Testing